backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.3
python-multipart==0.0.9
PyJWT==2.9.0
minio==7.2.7
structlog==24.4.0
mypy==1.11.2
ruff==0.6.4
black==24.8.0
